Experimental: Long stack trace support
node-mysql is definitley not the right place to fix this, but support
for this in the node core is long overdue and
tlrobinson/long-stack-traces is very hacky, so I think this is justified
for now, considering how little code is needed to make it work.

But please let me know if you have feedback on this.

Closes #198