Improve ChangeUser implementation and test
The previous implementation did not work correctly when calling
changeUser() before the handshake was done which became apparent when
restructuring the test. The new implemenetation passes the
handshakeInitializationPacket to the ChangeUser#start method, so it does
not have to be available when enqueueing the sequence.

Additionally I removed some uneeded code, such as
ChangeUser#scrambleBuffer and convention violations such as accessing
private properties of another object.