Switch to UTF8_GENERAL_CI as default charset
Also make the charset configurable. This fixes #200 and aligns
node-mysql with most other drivers as far as the default charset is
concerned.

As for #200, it seems that there is a MySQL bug that will cause certain
queries to run massively slower when using UTF8_UNICODE_CI, even if the
conditions used should not involve any string operations. I will file a
bug for that as soon as I find a little more time (I'm at a conference
right now).