Fix parsing DATE strings and the test for it
JS apparently parses date strings as if they were in GMT if they do not
include a time component, but parses them as if they were in local time
otherwise. We always want to treat dates as if they were local time, so
this patch fixes the problem, and also adjusts the test to not run into
it itself.