docs: remove handleDisconnect example
This removes the code example that tries to create a new connection
after a connection is disconnected. This example came about before
there was a `Pool` in this library and was necessary. Now this example
is simply mis-leading and users end up trying to copy-and-paste it
into a stand-alone file and get confused on how to replace the
`connection` variable reference. `Pool` is the official solution.